make _columns mutable to change them when coloning table

+++ exposed-tests/src/main/kotlin/org/jetbrains/exposed/sql/tests/DatabaseTestsBase.kt
@@ -5,6 +5,7 @@ import org.h2.engine.Mode
 import org.jetbrains.exposed.sql.*
 import org.jetbrains.exposed.sql.transactions.transaction
 import org.jetbrains.exposed.sql.transactions.transactionManager
+import org.jetbrains.exposed.sql.vendors.currentDialect
 import org.testcontainers.containers.MySQLContainer
 import java.sql.Connection
 import java.util.*
@@ -162,14 +163,17 @@ abstract class DatabaseTestsBase {
     fun withSchemas (excludeSettings: List<TestDB>, vararg schemas: Schema, statement: Transaction.() -> Unit) {
         (TestDB.enabledInTests() - excludeSettings).forEach { testDB ->
             withDb(testDB) {
-                SchemaUtils.createSchema(*schemas)
-                try {
-                    statement()
-                    commit() // Need commit to persist data before drop schemas
-                } finally {
-                    val cascade = it != TestDB.SQLSERVER
-                    SchemaUtils.dropSchema(*schemas, cascade = cascade)
-                    commit()
+                /** Testing only dialects that supports schema. */
+                if (currentDialect.supportsCreateSchema) {
+                    SchemaUtils.createSchema(*schemas)
+                    try {
+                        statement()
+                        commit() // Need commit to persist data before drop schemas
+                    } finally {
+                        val cascade = it != TestDB.SQLSERVER
+                        SchemaUtils.dropSchema(*schemas, cascade = cascade)
+                        commit()
+                    }
                 }
             }
         }
